requirements.txt
opencv-python==4.8.0.76
pytesseract==0.3.10
spacy==3.6.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
psycopg2-binary==2.9.6
flask==2.3.3
flask-sqlalchemy==3.0.5
pillow==10.0.0
pdf2image==1.16.3
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
# Additional dependencies for enhanced image processing
imutils==0.5.4
scikit-image==0.21.0
poppler-utils==0.1.0
